move files to different package
