+++ annotation-processor/src/main/java/me/bristermitten/mittenlib/annotations/compile/ConfigImplGenerator.java
@@ -4,7 +4,6 @@ import com.squareup.javapoet.*;
 import me.bristermitten.mittenlib.annotations.ast.AbstractConfigStructure;
 import me.bristermitten.mittenlib.annotations.ast.ConfigTypeSource;
 import me.bristermitten.mittenlib.annotations.ast.Property;
-import me.bristermitten.mittenlib.annotations.config.*;
 import me.bristermitten.mittenlib.config.GeneratedConfig;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
