+++ annotation-processor/src/main/java/me/bristermitten/mittenlib/annotations/compile/FieldClassNameGenerator.java
@@ -1,4 +1,4 @@
-package me.bristermitten.mittenlib.annotations.config;
+package me.bristermitten.mittenlib.annotations.compile;
 
 import me.bristermitten.mittenlib.annotations.ast.Property;
 import me.bristermitten.mittenlib.annotations.util.TypesUtil;
