+++ annotation-processor/src/main/java/me/bristermitten/mittenlib/annotations/compile/MethodNames.java
@@ -1,4 +1,4 @@
-package me.bristermitten.mittenlib.annotations.config;
+package me.bristermitten.mittenlib.annotations.compile;
 
 import com.google.inject.Singleton;
 import me.bristermitten.mittenlib.annotations.ast.Property;
