+++ annotation-processor/src/main/java/me/bristermitten/mittenlib/annotations/parser/ConfigClassParser.java
@@ -6,7 +6,7 @@ import io.toolisticon.aptk.tools.corematcher.AptkCoreMatchers;
 import io.toolisticon.aptk.tools.wrapper.ElementWrapper;
 import io.toolisticon.aptk.tools.wrapper.TypeElementWrapper;
 import me.bristermitten.mittenlib.annotations.ast.*;
-import me.bristermitten.mittenlib.annotations.config.ConfigNameCache;
+import me.bristermitten.mittenlib.annotations.compile.ConfigNameCache;
 import me.bristermitten.mittenlib.annotations.util.ElementsFinder;
 import me.bristermitten.mittenlib.annotations.util.TypesUtil;
 import me.bristermitten.mittenlib.config.ConfigUnion;
