+++ annotation-processor/src/test/java/me/bristermitten/mittenlib/annotations/config/FieldClassNameGeneratorTest.java
@@ -2,6 +2,7 @@ package me.bristermitten.mittenlib.annotations.config;
 
 import com.google.testing.compile.Compilation;
 import com.google.testing.compile.JavaFileObjects;
+import me.bristermitten.mittenlib.annotations.compile.ConfigProcessor;
 import me.bristermitten.mittenlib.config.names.NamingPatterns;
 import org.junit.jupiter.api.Test;
